+++ src/Steps/Refiners/String/StrAfterLast.php
@@ -11,7 +11,7 @@ class StrAfterLast extends AbstractRefiner
     public function refine(mixed $value): mixed
     {
         if (!is_string($value)) {
-            $this->logTypeWarning('Str::afterLast()', $value);
+            $this->logTypeWarning('StringRefiner::afterLast()', $value);
 
             return $value;
         }