+++ tests/Steps/Refiners/String/AfterLastTest.php
@@ -15,7 +15,7 @@ it('logs a warning and returns the unchanged value when $value is not of type st
 
     $logOutput = $this->getActualOutputForAssertion();
 
-    expect($logOutput)->toContain('Refiner Str::afterLast() can\'t be applied to value of type ' . gettype($value));
+    expect($logOutput)->toContain('Refiner StringRefiner::afterLast() can\'t be applied to value of type ' . gettype($value));
 
     expect($refinedValue)->toBe($value);
 })->with([